% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_environ.R
\name{get_environ}
\alias{get_environ}
\title{get_environ}
\usage{
get_environ(base_folder = "environ_files")
}
\arguments{
\item{base_folder}{A folder to create if it doesn't exist to store environ data}
}
\description{
get_environ
}
