% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_lazy_update.R
\docType{data}
\name{git_lazy_update_dev}
\alias{git_lazy_update_dev}
\title{git_lazy_update_dev}
\format{
An object of class \code{character} of length 1.
}
\usage{
git_lazy_update_dev
}
\description{
git_lazy_update_dev
}
\keyword{datasets}
