% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_glue_message.R
\name{ndex_glue_message}
\alias{ndex_glue_message}
\title{ndex_glue_me}
\usage{
ndex_glue_message(string, ...)
}
\arguments{
\item{string}{A string with {} syntax}

\item{...}{Additional parameters to glue}
}
\description{
ndex_glue_me
}
