% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndex_glue_system.R
\name{ndex_glue_system}
\alias{ndex_glue_system}
\title{ndex_glue_system}
\usage{
ndex_glue_system(string, ...)
}
\arguments{
\item{string}{A string with {} syntax}

\item{...}{Additional parameters to glue}
}
\description{
ndex_glue_system
}
